chore: remove `ccdm` from URLs
rename `ccdm` into `typescript`
Fixes #928
