+++ documentation/Overview.asciidoc
@@ -106,14 +106,14 @@ The Vaadin Flow documentation is arranged in the following sections and we recom
 * <<polymer-templates/tutorial-template-model-encoders#,Using Model Encoders with a PolymerTemplate Model>>
 
 == Creating UI in TypeScript
-* <<ccdm/quick-start-guide-v15#, Start a new V15 project>>
-* <<ccdm/quick-start-guide-v14#, Migrate a V14 project>>
-* <<ccdm/client-side-bootstrapping#, Client-side Bootstrapping>>
-* <<ccdm/client-side-routing#, Client-side Routing>>
-* <<ccdm/intro-to-typescript-in-v15#, Intro to TypeScript in Vaadin 15>>
-* <<ccdm/how-to-create-api-endpoint#, How to Create an API Endpoint>>
-* <<ccdm/how-to-access-backend-from-typescript#, How to Access the Backend from TypeScript>>
-* <<ccdm/client-side-security#, Security in client-side>>
+* <<typescript/quick-start-guide-v15#, Start a new V15 project>>
+* <<typescript/quick-start-guide-v14#, Migrate a V14 project>>
+* <<typescript/client-side-bootstrapping#, Client-side Bootstrapping>>
+* <<typescript/client-side-routing#, Client-side Routing>>
+* <<typescript/intro-to-typescript-in-v15#, Intro to TypeScript in Vaadin 15>>
+* <<typescript/how-to-create-api-endpoint#, How to Create an API Endpoint>>
+* <<typescript/how-to-access-backend-from-typescript#, How to Access the Backend from TypeScript>>
+* <<typescript/client-side-security#, Security in client-side>>
 
 == Using Vaadin with Spring
 * <<spring/tutorial-spring-basic#,Use Vaadin with Spring>>
