+++ documentation/src/main/html/typescript/ClientSideBootstrapping.html
@@ -1,4 +1,4 @@
-tutorial::ccdm/client-side-bootstrapping.asciidoc
+tutorial::typescript/client-side-bootstrapping.asciidoc
 <!DOCTYPE html>
 <html lang="en">
 <head>
