+++ documentation/src/main/html/typescript/ClientSideRouting.html
@@ -1,4 +1,4 @@
-tutorial::ccdm/client-side-routing.asciidoc
+tutorial::typescript/client-side-routing.asciidoc
 <html>
   <body>
     <h1>My App</h1>
