+++ documentation/src/main/html/typescript/QuickStart.html
@@ -1,4 +1,4 @@
-tutorial::ccdm/quick-start-guide.asciidoc
+tutorial::typescript/quick-start-guide.asciidoc
 <html>
   <body>
     <div id="outlet"></div>
