+++ documentation/src/main/html/typescript/index-for-typescript-1.html
@@ -1,4 +1,4 @@
-tutorial::ccdm/intro-to-typescript-in-v15.asciidoc
+tutorial::typescript/intro-to-typescript-in-v15.asciidoc
 
 <!DOCTYPE html>
 <html lang="en">
