+++ documentation/src/main/java/com/vaadin/flow/tutorial/typescript/TypeConversion.java
@@ -13,14 +13,14 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.tutorial.ccdm;
+package com.vaadin.flow.tutorial.typescript;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonSetter;
 
 import com.vaadin.flow.tutorial.annotations.CodeFor;
 
-@CodeFor("ccdm/type-conversion.asciidoc")
+@CodeFor("typescript/type-conversion.asciidoc")
 public class TypeConversion {
     public enum TestEnum {
