+++ documentation/typescript/client-side-bootstrapping.asciidoc
