+++ documentation/typescript/client-side-security.asciidoc
