+++ documentation/typescript/codecompletion.gif
