+++ documentation/typescript/how-to-access-backend-from-typescript.asciidoc
