+++ documentation/typescript/how-to-create-api-endpoint.asciidoc
